Updated GameClock to handle delays. Changed group of auto. Moved XbotTelemetry to robotcore/utils. More logging.
