+++ Team3231/src/main/java/org/xbot/ftc/operatingcode/autonomous/jewel_smacker/BlueAuto.java
@@ -3,11 +3,9 @@ package org.xbot.ftc.operatingcode.autonomous.jewel_smacker;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.xbot.ftc.robotcore.subsystems.arm.JewelArm;
-import org.xbot.ftc.robotcore.subsystems.drive.Drive;
 import org.xbot.ftc.robotcore.subsystems.vision.XbotColorSensor;
 
-@Autonomous(name="Blue", group="Jewel")
+@Autonomous(name="Jewel: Blue", group="Test")
 public class BlueAuto extends LinearOpMode {
 
     @Override
