+++ Team3231/src/main/java/org/xbot/ftc/operatingcode/autonomous/jewel_smacker/RedAuto.java
@@ -3,11 +3,9 @@ package org.xbot.ftc.operatingcode.autonomous.jewel_smacker;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.xbot.ftc.robotcore.subsystems.arm.JewelArm;
-import org.xbot.ftc.robotcore.subsystems.drive.Drive;
 import org.xbot.ftc.robotcore.subsystems.vision.XbotColorSensor;
 
-@Autonomous(name="Red", group="Jewel")
+@Autonomous(name="Jewel: Red", group="Test")
 public class RedAuto extends LinearOpMode {
 
     @Override
